refactor: extract metrics logic out of push_handler
